modules/person_kafka_consumer/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements/production.in
#
greenlet==1.1.2
    # via sqlalchemy
kafka-python==2.0.2
    # via -r requirements\common.in
marshmallow==3.14.0
    # via marshmallow-sqlalchemy
marshmallow-sqlalchemy==0.26.1
    # via -r requirements\common.in
psycopg2-binary==2.9.1
    # via -r requirements\common.in
python-dotenv==0.19.1
    # via -r requirements\common.in
sqlalchemy==1.4.26
    # via
    #   -r requirements\common.in
    #   marshmallow-sqlalchemy
werkzeug==2.0.2
    # via -r requirements\common.in
